Resolved issues with original sample app. 1) Memory leaks because delete was never called on the images containing a barcode. Both in ImageVideoDlg.cpp and in ScanToConnect.Dlg.cpp 2) Memory leak when the ScannerEventSink was being released. The pointer was assigned a NULL value but the memory itself was never deleted. That may be fine in C#, but not in C++. Also, the variable was assigned NULL (correct) and 0 (incorrect and redundant). 3) DWORD values were getting implicitly cast to ints. While this was acceptable since the values would always be within an int range, I added static_cast<int>() around the values to suppress the warnings. 4) Soon to be deprecated /Gm compile switch was removed. Multiprocessor build MP was enabled in its place. Unless a platform older than VS 2015 is to be used, this should be backward compatible.

+++ SampleApp_CPP/RtaList.cpp
@@ -301,17 +301,17 @@ void CRtaListCtrl::OnCustomDraw(NMHDR* pNMHDR, LRESULT* pResult)
 
     case CDDS_ITEMPREPAINT | CDDS_SUBITEM:
         // Check if the current subitem is in a checkbox column
-        if (IsCheckboxColumn(pLVCD->nmcd.dwItemSpec, pLVCD->iSubItem))
+        if (IsCheckboxColumn(static_cast<int>(pLVCD->nmcd.dwItemSpec), pLVCD->iSubItem))
         {
             // Get a device context (DC) from the custom draw structure
             CDC* pDC = CDC::FromHandle(pLVCD->nmcd.hdc);
             CRect rect;
 
             // Get the rectangle area of the subitem
-            GetSubItemRect(pLVCD->nmcd.dwItemSpec, pLVCD->iSubItem, LVIR_LABEL, rect);
+            GetSubItemRect(static_cast<int>(pLVCD->nmcd.dwItemSpec), pLVCD->iSubItem, LVIR_LABEL, rect);
 
             // Draw the checkbox in the subitem's rectangle
-            DrawCheckbox(pDC, this, pLVCD->nmcd.dwItemSpec, pLVCD->iSubItem, rect);
+            DrawCheckbox(pDC, this, static_cast<int>(pLVCD->nmcd.dwItemSpec), pLVCD->iSubItem, rect);
 
             // Skip the default drawing for this subitem (since we've custom drawn it)
             *pResult = CDRF_SKIPDEFAULT;
